% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_factors.R
\name{add_factors_deprecated}
\alias{add_factors_deprecated}
\title{Include add factors}
\usage{
add_factors_deprecated(df, last_date)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Include add factors
}
