% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_social_benefits_components.R
\name{add_social_benefit_components}
\alias{add_social_benefit_components}
\title{Title}
\usage{
add_social_benefit_components(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Title
}
