% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate federal and state components}
\usage{
aggregate(df, total, federal, state)
}
\arguments{
\item{df}{}

\item{total}{}

\item{federal}{}

\item{state}{}
}
\value{

}
\description{
Aggregate federal and state components
}
\examples{
projections \%>\%
   aggregate(gtfp, gftfp, gstfp) 
}
