% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{alternative_tax_scenario}
\alias{alternative_tax_scenario}
\title{Alternative tax scenario}
\usage{
alternative_tax_scenario(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Alternative tax scenario
}
