% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_to_quarter.R
\name{annual_to_quarter}
\alias{annual_to_quarter}
\title{Convert annual data to quarterly}
\usage{
annual_to_quarter(df)
}
\arguments{
\item{df}{}

\item{var}{}
}
\value{

}
\description{
Convert annual data to quarterly
}
