% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{bare_to_name}
\alias{bare_to_name}
\title{Quosure to a name: quo_name}
\usage{
bare_to_name(x, nm)
}
\arguments{
\item{x}{}

\item{nm}{}
}
\value{

}
\description{
Quosure to a name: quo_name
}
\examples{
bare_to_name(mtcars, quo(this_is_42))
}
