% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{bare_to_quo}
\alias{bare_to_quo}
\title{Bare to quosure: quo}
\usage{
bare_to_quo(x, var)
}
\arguments{
\item{x}{}

\item{var}{}
}
\value{

}
\description{
Bare to quosure: quo
}
\examples{
bare_to_quo(mtcars, quo(cyl))
}
