% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{bare_to_quo_in_func}
\alias{bare_to_quo_in_func}
\title{Bare to quosure in function: enquo}
\usage{
bare_to_quo_in_func(x, var)
}
\arguments{
\item{x}{}

\item{var}{}
}
\value{

}
\description{
Bare to quosure in function: enquo
}
\examples{
bare_to_quo_in_func(mtcars, mpg)
}
