% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{bare_to_quo_mult}
\alias{bare_to_quo_mult}
\title{Multiple bares to quosures:quos}
\usage{
bare_to_quo_mult(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Multiple bares to quosures:quos
}
