% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{bare_to_quo_mult_chars}
\alias{bare_to_quo_mult_chars}
\title{Multiple characters to names: syms}
\usage{
bare_to_quo_mult_chars(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Multiple characters to names: syms
}
