% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-projections.R
\docType{data}
\name{budg}
\alias{budg}
\title{CBO 10-Year Budget Projections}
\format{
A data frame with 56 rows and 10 variables:
\describe{
\item{\code{fy}}{double. DESCRIPTION.}
\item{\code{gftfp}}{double. DESCRIPTION.}
\item{\code{gfrpt}}{double. DESCRIPTION.}
\item{\code{gfrpri}}{double. DESCRIPTION.}
\item{\code{gfrcp}}{double. DESCRIPTION.}
\item{\code{gfrs}}{double. DESCRIPTION.}
\item{\code{yptmr}}{double. DESCRIPTION.}
\item{\code{yptmd}}{double. DESCRIPTION.}
\item{\code{yptu}}{double. DESCRIPTION.}
\item{\code{date}}{double. DESCRIPTION.}
}
}
\usage{
budg
}
\description{
Projections of spending and revenues by category and of deficits and debt held by the public.
}
\keyword{datasets}
