% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumption.R
\name{calculate_consumption}
\alias{calculate_consumption}
\title{Calculate consumption's
Create consumption}
\usage{
calculate_consumption(df, taxes_transfers)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Calculate consumption's
Create consumption
}
