% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{calculate_mpc}
\alias{calculate_mpc}
\title{Apply mpcs to taxes and transfers}
\usage{
calculate_mpc(df, taxes_transfers)
}
\arguments{
\item{df}{}

\item{taxes_transfers}{}
}
\value{

}
\description{
Apply mpcs to taxes and transfers
}
