% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{char_to_quo}
\alias{char_to_quo}
\title{character to name: sym}
\usage{
char_to_quo(x, var)
}
\arguments{
\item{x}{}

\item{var}{}
}
\value{

}
\description{
character to name: sym
}
\examples{
char_to_quo(mtcars, "vs")
}
