% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_forecast.R
\name{coalesce_growth}
\alias{coalesce_growth}
\title{Coalesce missing observations with projected growth rate}
\usage{
coalesce_growth(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{

}
\description{
Coalesce missing observations with projected growth rate
}
