% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_helpers.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Compare between versions
Helper functions to visualize differences between different versions of
the FIM. By default this returns a comparison between the previous months total
fiscal impact and the current one.}
\usage{
comparison_plot(df = old, df_new = new, variable = fiscal_impact, title = "")
}
\arguments{
\item{df}{}

\item{df_new}{}

\item{variable}{}

\item{title}{}
}
\value{

}
\description{
Compare between versions
Helper functions to visualize differences between different versions of
the FIM. By default this returns a comparison between the previous months total
fiscal impact and the current one.
}
