% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumption.R, R/mpc.R
\name{consumption}
\alias{consumption}
\alias{consumption_social_benefits}
\alias{consumption_coronavirus_relief_fund}
\alias{consumption_rebate_checks}
\alias{consumption_health_outlays}
\alias{consumption_corporate_taxes}
\alias{consumption_non_corporate_taxes}
\alias{consumption_subsidies}
\alias{consumption_subsidies_second_draw}
\alias{consumption_subsidies_rra}
\alias{consumption_ui}
\alias{mpc}
\alias{mpc_social_benefits}
\alias{mpc_coronavirus_relief_fund}
\alias{mpc_rebate_checks}
\alias{mpc_health_outlays}
\alias{mpc_corporate_taxes}
\alias{mpc_non_corporate_taxes}
\alias{mpc_subsidies}
\alias{mpc_subsidies_second_draw}
\alias{mpc_subsidies_rra}
\alias{mpc_ui}
\alias{mpc_vulnerable_arp}
\alias{mpc_direct_aid_arp}
\alias{mpc_small_businesses_arp}
\alias{mpc_ui_arp}
\alias{mpc_non_health_grants_arp}
\title{Get dollars of consumptions}
\usage{
consumption(mpc = 1, timing)

consumption_social_benefits(x)

consumption_coronavirus_relief_fund(x)

consumption_rebate_checks(x)

consumption_health_outlays(x)

consumption_corporate_taxes(x)

consumption_non_corporate_taxes(x)

consumption_subsidies(x)

consumption_subsidies_second_draw(x)

consumption_subsidies_rra(x)

consumption_ui(x)

mpc(mpc = 1, timing)

mpc_social_benefits(x)

mpc_coronavirus_relief_fund(x)

mpc_rebate_checks(x)

mpc_health_outlays(x)

mpc_corporate_taxes(x)

mpc_non_corporate_taxes(x)

mpc_subsidies(x)

mpc_subsidies_second_draw(x)

mpc_subsidies_rra(x)

mpc_ui(x)

mpc_vulnerable_arp(x)

mpc_direct_aid_arp(x)

mpc_small_businesses_arp(x)

mpc_ui_arp(x)

mpc_non_health_grants_arp(x)
}
\arguments{
\item{mpc}{numeric}

\item{timing}{numeric}
}
\description{
Get dollars of consumptions

Get dollars of consumptions
In order to capture the effect of taxes and transfers to GDP we need to translate government expenditures into dollars of consumption.
}
