% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{continuous_summary}
\alias{continuous_summary}
\title{Continuous summary
Quickly summarize variable count, mean, standard deviation, median, min, and max.}
\usage{
continuous_summary(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{

}
\description{
Continuous summary
Quickly summarize variable count, mean, standard deviation, median, min, and max.
}
