% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R, R/contributions.R,
%   R/equations.R
\name{contribution}
\alias{contribution}
\title{FUNCTION_TITLE}
\usage{
contribution(var, deflator)

contribution(var, deflator)

contribution(var, deflator)
}
\arguments{
\item{var}{}

\item{deflator}{}

\item{.data}{}

\item{...}{}
}
\value{
Contribution of var to real gdp growth




}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{as_string}},\code{\link[rlang]{sym}}
}
