% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions_purchases_grants}
\alias{contributions_purchases_grants}
\title{Calculate contributions of grants and purchases}
\usage{
contributions_purchases_grants(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Calculate contributions of grants and purchases
}
