% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R, R/equations.R
\name{counterfactual}
\alias{counterfactual}
\title{Counterfactual spending}
\usage{
counterfactual(var, deflator)

counterfactual(var, deflator)
}
\arguments{
\item{var}{}

\item{deflator}{}

\item{.data}{}

\item{...}{}
}
\value{



}
\description{
Counterfactual spending

Title
}
