% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_assumptions.R
\name{deflators_growth}
\alias{deflators_growth}
\title{Deflators growth rate assumptions}
\usage{
deflators_growth(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Deflators growth rate assumptions
}
