% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic-projections-data.R
\docType{data}
\name{econ}
\alias{econ}
\title{CBO Quarterly Economic projections}
\format{
A data frame with 56 rows and 16 variables:
\describe{
\item{\code{date}}{double. DESCRIPTION.}
\item{\code{gdp}}{double. DESCRIPTION.}
\item{\code{gdph}}{double. DESCRIPTION.}
\item{\code{gdppothq}}{double. DESCRIPTION.}
\item{\code{gdppotq}}{double. DESCRIPTION.}
\item{\code{dc}}{double. DESCRIPTION.}
\item{\code{jgdp}}{double. DESCRIPTION.}
\item{\code{c}}{double. DESCRIPTION.}
\item{\code{ch}}{double. DESCRIPTION.}
\item{\code{gh}}{double. DESCRIPTION.}
\item{\code{gfh}}{double. DESCRIPTION.}
\item{\code{gsh}}{double. DESCRIPTION.}
\item{\code{g}}{double. DESCRIPTION.}
\item{\code{gf}}{double. DESCRIPTION.}
\item{\code{gs}}{double. DESCRIPTION.}
\item{\code{cpiu}}{double. DESCRIPTION.}
}
}
\usage{
econ
}
\description{
Projections of output, prices, labor market measures,
interest rates, and income. Beginning in January 2020, these files also include projections of
potential GDP (the economy’s maximum sustainable output) and its underlying inputs
}
\keyword{datasets}
