% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{federal_transfers_growth_rates}
\alias{federal_transfers_growth_rates}
\title{Get growth rates of federal transfers}
\usage{
federal_transfers_growth_rates(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Get growth rates of federal transfers
}
