% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{filter_func}
\alias{filter_func}
\title{Quoting full expression}
\usage{
filter_func(x, filter_exp)
}
\arguments{
\item{x}{}

\item{filter_exp}{}
}
\value{

}
\description{
Quoting full expression
}
