% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim-package.R
\docType{package}
\name{fim-package}
\alias{fim}
\alias{fim-package}
\title{fim: Fiscal Impact Measure}
\description{
The Hutchins Center Fiscal Impact Measure shows how much local, state, and federal tax and spending policy adds to or subtracts from overall economic growth, and provides a near-term forecast of fiscal policies’ effects on economic activity.
}
\author{
\strong{Maintainer}: Manuel Alcala Kovalski \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
