% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_figures.R
\docType{data}
\name{fim_colors}
\alias{fim_colors}
\title{Set colors for components}
\format{
An object of class \code{character} of length 4.
}
\usage{
fim_colors
}
\value{

}
\description{
Set colors for components
}
\keyword{datasets}
