% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{fim_cols}
\alias{fim_cols}
\title{Function to extract fim colors as hex codes}
\usage{
fim_cols(...)
}
\arguments{
\item{...}{Character names of fim_colors}
}
\description{
Function to extract fim colors as hex codes
}
