% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{fim_pal}
\alias{fim_pal}
\title{Return function to interpolate a fim color palette}
\usage{
fim_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fim_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a fim color palette
}
