% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{fim_colors}
\alias{fim_colors}
\alias{fim_palettes}
\title{fim Institution palettes}
\format{
An object of class \code{character} of length 4.

An object of class \code{list} of length 2.
}
\source{
fim Style Guide
x
}
\usage{
fim_colors

fim_palettes
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the fim Style Guide.
x
}
\concept{fim palettes}
\keyword{datasets}
