% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_forecast.R
\name{forecast2}
\alias{forecast2}
\title{Forecast using projected growth rates}
\usage{
forecast2(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{

}
\description{
Forecast using projected growth rates
}
