% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{forecast_state_taxes}
\alias{forecast_state_taxes}
\title{Forecast state taxes}
\usage{
forecast_state_taxes(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Forecast state taxes
}
