% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_names.R
\name{get_components_names}
\alias{get_components_names}
\title{Components
Names of the components we project.}
\usage{
get_components_names()
}
\value{

}
\description{
Components
Names of the components we project.
}
