% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_previous_month}
\alias{get_previous_month}
\title{Previous month
Helper function to retrieve the previous months date.
This is particularly helpful when navigating across the
results and reports directories.}
\usage{
get_previous_month()
}
\value{

}
\description{
Previous month
Helper function to retrieve the previous months date.
This is particularly helpful when navigating across the
results and reports directories.
}
