% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_figures.R
\name{get_recession_shade}
\alias{get_recession_shade}
\title{Get recession start and end dates for shading}
\usage{
get_recession_shade(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Get recession start and end dates for shading
}
