% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_figures.R
\name{get_recessions}
\alias{get_recessions}
\title{Create recessions data frame
Data frame with start and end dates of U.S. recessions from NBER.}
\usage{
get_recessions(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Create recessions data frame
Data frame with start and end dates of U.S. recessions from NBER.
}
