% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haver.R
\docType{data}
\name{haver}
\alias{haver}
\title{Historical Data}
\format{
A data frame with 203 rows and 52 variables:
\describe{
\item{\code{date}}{double. Quarterly.}
\item{\code{gdp}}{double. Nominal gross domestic product.}
\item{\code{c}}{double. Nominal consumption.}
\item{\code{ch}}{double. Real consumption.}
\item{\code{gdph}}{double. Real Gross Domestic Product.}
\item{\code{jc}}{double. Consumption deflator.}
\item{\code{jgdp}}{double. DESCRIPTION.}
\item{\code{jgf}}{double. DESCRIPTION.}
\item{\code{jgs}}{double. DESCRIPTION.}
\item{\code{jgse}}{double. DESCRIPTION.}
\item{\code{jgsi}}{double. DESCRIPTION.}
\item{\code{ptgh}}{double. DESCRIPTION.}
\item{\code{ptgsh}}{double. DESCRIPTION.}
\item{\code{ptgfh}}{double. DESCRIPTION.}
\item{\code{yptmr}}{double. DESCRIPTION.}
\item{\code{yptmd}}{double. DESCRIPTION.}
\item{\code{yptu}}{double. DESCRIPTION.}
\item{\code{gtfp}}{double. DESCRIPTION.}
\item{\code{ypog}}{double. DESCRIPTION.}
\item{\code{yptx}}{double. DESCRIPTION.}
\item{\code{ytpi}}{double. DESCRIPTION.}
\item{\code{yctlg}}{double. DESCRIPTION.}
\item{\code{g}}{double. DESCRIPTION.}
\item{\code{grcsi}}{double. DESCRIPTION.}
\item{\code{dc}}{double. DESCRIPTION.}
\item{\code{gf}}{double. DESCRIPTION.}
\item{\code{gs}}{double. DESCRIPTION.}
\item{\code{gfh}}{double. DESCRIPTION.}
\item{\code{gsh}}{double. DESCRIPTION.}
\item{\code{gfrpt}}{double. DESCRIPTION.}
\item{\code{gfrpri}}{double. DESCRIPTION.}
\item{\code{gfrcp}}{double. DESCRIPTION.}
\item{\code{gfrs}}{double. DESCRIPTION.}
\item{\code{gftfp}}{double. DESCRIPTION.}
\item{\code{gfeg}}{double. DESCRIPTION.}
\item{\code{gsrpt}}{double. DESCRIPTION.}
\item{\code{gsrpri}}{double. DESCRIPTION.}
\item{\code{gsrcp}}{double. DESCRIPTION.}
\item{\code{gsrs}}{double. DESCRIPTION.}
\item{\code{gstfp}}{double. DESCRIPTION.}
\item{\code{gset}}{double. DESCRIPTION.}
\item{\code{gfeghhx}}{double. DESCRIPTION.}
\item{\code{gfeghdx}}{double. DESCRIPTION.}
\item{\code{gfeigx}}{double. DESCRIPTION.}
\item{\code{gfsub}}{double. DESCRIPTION.}
\item{\code{gssub}}{double. DESCRIPTION.}
\item{\code{gsub}}{double. DESCRIPTION.}
\item{\code{gftfbusx}}{double. DESCRIPTION.}
\item{\code{pcw}}{double. DESCRIPTION.}
\item{\code{gdppothq}}{double. DESCRIPTION.}
\item{\code{gdppotq}}{double. DESCRIPTION.}
\item{\code{recessq}}{double. DESCRIPTION.}
}
}
\usage{
haver
}
\description{
Historical quarterly national accounts and economic statistics data from Haver
Analytics. Data is from 1970-03-31 until present.
}
\keyword{datasets}
