% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_assumptions.R
\name{health_growth}
\alias{health_growth}
\title{Health outlays growth assumptions
Grow Federal Medicaid grants to states and Health & Hospital grants
with Medicaid}
\usage{
health_growth(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Health outlays growth assumptions
Grow Federal Medicaid grants to states and Health & Hospital grants
with Medicaid
}
