% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{health_outlays_growth_rates}
\alias{health_outlays_growth_rates}
\title{Get health growth rates}
\usage{
health_outlays_growth_rates(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Get health growth rates
}
