% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{implicit_price_deflators}
\alias{implicit_price_deflators}
\title{Implicit price deflators}
\usage{
implicit_price_deflators(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Implicit price deflators
}
