% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make_cumulative_growth_rates}
\alias{make_cumulative_growth_rates}
\title{Calculate cumulative growth rate}
\usage{
make_cumulative_growth_rates(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Calculate cumulative growth rate
}
