% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make_forecast}
\alias{make_forecast}
\title{Title}
\usage{
make_forecast(df, forecast_period, components)
}
\arguments{
\item{df}{}

\item{forecast_period}{}

\item{components}{}
}
\value{

}
\description{
Title
}
