% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocations.R
\name{medicaid_reallocation}
\alias{medicaid_reallocation}
\title{Medicaid reallocation
Before  constructing projections, we reallocate Federal Medicaid grants from
S&L governments to the Federal govt.}
\usage{
medicaid_reallocation(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Medicaid reallocation
Before  constructing projections, we reallocate Federal Medicaid grants from
S&L governments to the Federal govt.
}
