% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc_corporate_taxes.R
\name{mpc_corporate_taxes}
\alias{mpc_corporate_taxes}
\title{FUNCTION_TITLE}
\usage{
mpc_corporate_taxes(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
