% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc_health_outlays.R
\name{mpc_health_outlays}
\alias{mpc_health_outlays}
\title{FUNCTION_TITLE}
\usage{
mpc_health_outlays(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
