% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc_noncorp_taxes_CRN19.R
\name{mpc_noncorp_taxes_CRN19}
\alias{mpc_noncorp_taxes_CRN19}
\title{FUNCTION_TITLE}
\usage{
mpc_noncorp_taxes_CRN19(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
