% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_where.R
\name{mutate_where}
\alias{mutate_where}
\title{Mutate on rows using logical subsetting}
\usage{
mutate_where(.data, .where, ...)
}
\arguments{
\item{.data}{}

\item{.where}{}

\item{...}{}
}
\value{

}
\description{
Mutate on rows using logical subsetting
}
