% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haver.R
\docType{data}
\name{national_accounts}
\alias{national_accounts}
\title{Historical Data}
\format{
A data frame with 203 rows and 52 variables:
\describe{
\item{\code{date}}{Quarterly.}
\item{\code{gdp}}{Nominal Gross Domestic Product (SAAR, Bil.$).}
\item{\code{c}}{Personal Consumption Expenditures (SAAR, Bil.$).}
\item{\code{ch}}{Real Personal Consumption Expenditures}
\item{\code{gdph}}{Real Gross Domestic Product (SAAR, Bil.Chn.2012$)}
\item{\code{jc}}{Personal Consumption Expenditures: Chain Price Index (SA, 2012=100)}
\item{\code{jgdp}}{Gross Domestic Product: Chain Price Index (SA, 2012=100).}
\item{\code{jgf}}{Federal Govt Consumption & Gross Investment: Chn Price Index(SA, 2012=100)}
\item{\code{jgs}}{State & Local Govt Consumption/Gross Investment: Chn Price Index(SA, 2012=100)}
\item{\code{jgse}}{State & Local Govt Consumption Expenditures: Chain Price Index (SA, 2012=100)}
\item{\code{jgsi}}{State & Local Govt Gross Investment: Chain Price Index (SA, 2012=100)}
\item{\code{ptgh}}{Government Consumption & Investment: Contrib to Real GDP \% Change (SAAR, \%Pt)}
\item{\code{ptgsh}}{State & Local Govt Consumption/Investment: Contrib to Real GDP \% Chg (SAAR, \%Pt)}
\item{\code{ptgfh}}{Federal Govt Consumption & Investment: Contrib to Real GDP \% Change (SAAR, \%Pt)}
\item{\code{yptmr}}{Government Social Benefit Payments to Persons: Medicare (SAAR, Bil.$)}
\item{\code{yptmd}}{Government Social Benefit Payments to Persons: Medicaid (SAAR, Bil.$)}
\item{\code{yptu}}{Govt Transf to Persons: Unemployment Insurance Benefits (SAAR, Bil.$)}
\item{\code{gtfp}}{Government Social Benefit Payments to Persons (SAAR, Bil.$)}
\item{\code{ypog}}{Personal Current Transfer Payments to Government (SAAR, Bil.$)}
\item{\code{yptx}}{Personal Current Taxes (SAAR, Bil.$)}
\item{\code{ytpi}}{Government Tax Receipts on Production & Imports (SAAR, Bil.$)}
\item{\code{yctlg}}{Government Tax Receipts on Corporate Income (SAAR, Bil.$)}
\item{\code{g}}{Government Consumption Expenditures & Gross Investment (SAAR, Bil.$)}
\item{\code{grcsi}}{Contributions for Government Social Insurance (SAAR, Bil.$)}
\item{\code{dc}}{Personal Consumption Expenditures: Implicit Price Deflator (SA, 2012=100)}
\item{\code{gf}}{Federal Government Consumption & Gross Investment (SAAR, Bil.$)}
\item{\code{gs}}{State & Local Government Consumption & Gross Investment (SAAR, Bil.$)}
\item{\code{gfh}}{Real Federal Government Consumption & Gross Investment (SAAR, Bil.Chn.2012$)}
\item{\code{gsh}}{Real State & Local Govt Consumption & Gross Investment (SAAR, Bil.Chn.2012$)}
\item{\code{gfrpt}}{Federal Government Personal Current Tax Receipts (SAAR, Bil.$)}
\item{\code{gfrpri}}{Federal Government Tax Receipts on Production & Imports (SAAR, Bil.$)}
\item{\code{gfrcp}}{Federal Government Tax Receipts on Corporate Income (SAAR, Bil.$)}
\item{\code{gfrs}}{Contributions for Federal Government Social Insurance (SAAR, Bil.$)}
\item{\code{gftfp}}{Federal Government Social Benefit Payments to Persons (SAAR, Bil.$)}
\item{\code{gfeg}}{Federal Grants-in-Aid to State & Local Governments (SAAR, Bil.$)}
\item{\code{gsrpt}}{State & Local Government Personal Current Tax Receipts (SAAR, Bil.$)}
\item{\code{gsrpri}}{State & Local Government Tax Receipts on Production & Imports (SAAR, Bil.$)}
\item{\code{gsrcp}}{State & Local Government Tax Receipts on Corporate Income (SAAR, Bil.$)}
\item{\code{gsrs}}{Contributions for State & Local Government Social Insurance (SAAR, Bil.$)}
\item{\code{gstfp}}{State & Local Government Social Benefit Payments to Persons (SAAR, Bil.$)}
\item{\code{gset}}{State & Local Government Expenditures (SAAR, Bil.$)}
\item{\code{gfeghhx}}{Federal Grants to State & Local Governments: Health & Hospitals (SAAR, Mil.$)}
\item{\code{gfeghdx}}{Federal Grants to State & Local Governments: Medicaid (SAAR, Mil.$)}
\item{\code{gfeigx}}{Federal Investment Grants to State and Local Governments (SAAR, Mil.$)}
\item{\code{gfsub}}{Federal Government Subsidies (SAAR, Bil.$)}
\item{\code{gssub}}{State & Local Government Subsidies (SAAR, Bil.$)}
\item{\code{gsub}}{Government Subsidies (SAAR, Bil.$)}
\item{\code{gftfbusx}}{Fed Transfer Payments/Persons: State Unemployment Insur Benefits (SAAR, Mil.$)}
\item{\code{pcw}}{CPI-W: All Items (SA, 1982-84=100), Bureau of Labor Statistics}
\item{\code{gdppothq}}{Real Potential Gross Domestic Product \link{CBO} (SAAR, Bil.Chn.2012$)}
\item{\code{gdppotq}}{Potential Gross Domestic Product \link{CBO} (SAAR, Bil.$)}
\item{\code{recessq}}{Quarterly NBER Recession/Expansion: Recession Shading}
}
}
\usage{
national_accounts
}
\description{
Historical quarterly national accounts and economic statistics data from Haver
Analytics. Data is from 1970-03-31 until present.
}
\keyword{datasets}
