% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_xlsx.R
\name{output_xlsx}
\alias{output_xlsx}
\title{Output xlsx}
\usage{
output_xlsx(data, names)
}
\arguments{
\item{data}{list of data frame objects to iterate over}

\item{names}{list of names for data frames}
}
\value{
Export excel workbooks
}
\description{
Export multiple xlsx files
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
