% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_interactive}
\alias{prepare_interactive}
\title{Prepare data for fim web interactive}
\usage{
prepare_interactive(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Prepare data for fim web interactive
}
