% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_forecast.R
\name{project}
\alias{project}
\title{Project series  with a growth rate}
\usage{
project(.data, ..., with = NULL, from = NULL)
}
\arguments{
\item{.data}{}

\item{...}{}

\item{with}{}
}
\value{

}
\description{
Project series  with a growth rate
}
