% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-projections.R,
%   R/economic-projections-data.R
\docType{data}
\name{projections}
\alias{projections}
\title{CBO 10-Year Budget Projections}
\format{
A data frame with 56 rows and 10 variables:
\describe{
\item{\code{date}}{Quarterly calendar data. Data was originally annual in fiscal years.}
\item{\code{gftfp}}{Federal Government Social Benefit Payments to Persons}
\item{\code{gfrpt}}{Federal Government Personal Current Tax Receipts}
\item{\code{gfrpri}}{Federal Government Tax Receipts on Production & Imports}
\item{\code{gfrcp}}{Federal Government Tax Receipts on Corporate Income}
\item{\code{gfrs}}{Contributions for Federal Government Social Insurance}
\item{\code{yptmr}}{Government Social Benefit Payments to Persons: Medicare}
\item{\code{yptmd}}{Government Social Benefit Payments to Persons: Medicaid}
\item{\code{yptu}}{Govt Transf to Persons: Unemployment Insurance Benefits}
}

A data frame with 56 rows and 16 variables:
\describe{
\item{\code{date}}{double. DESCRIPTION.}
\item{\code{gdp}}{double. DESCRIPTION.}
\item{\code{gdph}}{double. DESCRIPTION.}
\item{\code{gdppothq}}{double. DESCRIPTION.}
\item{\code{gdppotq}}{double. DESCRIPTION.}
\item{\code{dc}}{double. DESCRIPTION.}
\item{\code{jgdp}}{double. DESCRIPTION.}
\item{\code{c}}{double. DESCRIPTION.}
\item{\code{ch}}{double. DESCRIPTION.}
\item{\code{gh}}{double. DESCRIPTION.}
\item{\code{gfh}}{double. DESCRIPTION.}
\item{\code{gsh}}{double. DESCRIPTION.}
\item{\code{g}}{double. DESCRIPTION.}
\item{\code{gf}}{double. DESCRIPTION.}
\item{\code{gs}}{double. DESCRIPTION.}
\item{\code{cpiu}}{double. DESCRIPTION.}
}
}
\usage{
projections

projections
}
\description{
Projections of spending and revenues by category and of deficits and debt held by the public.

Projections of output, prices, labor market measures,
interest rates, and income. Beginning in January 2020, these files also include projections of
potential GDP (the economy’s maximum sustainable output) and its underlying inputs
}
\keyword{datasets}
