% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pull_data.R, R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{FUNCTION_TITLE}
\usage{
pull_data(series, database, start.date = "1970-01-01", frequency = "quarterly")

pull_data(series, database, start.date = "1970-01-01", frequency = "quarterly")
}
\arguments{
\item{series}{names of variables to pull}

\item{database}{name of database. FIM uses USNA and USECON}

\item{start.date}{PARAM_DESCRIPTION, Default: '1970-01-01'}

\item{frequency}{PARAM_DESCRIPTION, Default: 'quarterly'}
}
\value{
OUTPUT_DESCRIPTION

Dataframe
}
\description{
FUNCTION_DESCRIPTION

Wrapper function to pull data from Haver Analytics.
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
