% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchases_contribution2.R
\name{purchases_contributions2}
\alias{purchases_contributions2}
\title{Title}
\usage{
purchases_contributions2(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{

}
\description{
Title
}
