% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{q_a}
\alias{q_a}
\title{Quarterly annualized growth rate}
\usage{
q_a(x)
}
\arguments{
\item{x}{A quarterly time series object}
}
\value{
Annualized quarterly growth
}
\description{
Calculate quarterly annualized growth rates
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
