% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_g.R
\name{q_g}
\alias{q_g}
\title{Quarter over quarter growth rate}
\usage{
q_g(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculates quarter over quarter growth rate.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
