% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{q_g_fourthlag}
\alias{q_g_fourthlag}
\title{Q4 over Q4 growth rate}
\usage{
q_g_fourthlag(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculates 4 quarter lag growth rate.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
