% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{quo_to_text}
\alias{quo_to_text}
\title{Quosure to text: quo_text}
\usage{
quo_to_text(x, var)
}
\arguments{
\item{x}{}

\item{var}{}
}
\value{

}
\description{
Quosure to text: quo_text
}
\examples{
quo_to_text(mtcars, cyl)
}
