% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mpc.R
\name{read_mpc_file}
\alias{read_mpc_file}
\title{Title}
\usage{
read_mpc_file(path = "data/forecast.xlsx", sheet = "mpc", range = "B2:N14")
}
\arguments{
\item{path}{}

\item{sheet}{}

\item{range}{}
}
\value{

}
\description{
Title
}
