% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{remove_social_benefit_components}
\alias{remove_social_benefit_components}
\title{Separate unemployment insurance and rebate checks from social benefits}
\usage{
remove_social_benefit_components(df, component)
}
\arguments{
\item{df}{}

\item{component}{}
}
\value{

}
\description{
Separate unemployment insurance and rebate checks from social benefits
}
