% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumption.R
\name{rename_consumption}
\alias{rename_consumption}
\title{Rename consumption
Ad hoc function to rename column from 'minus_neutral' to 'post_consumption'}
\usage{
rename_consumption(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Rename consumption
Ad hoc function to rename column from 'minus_neutral' to 'post_consumption'
}
