% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{rename_mpc}
\alias{rename_mpc}
\title{Rename mpc
Ad hoc function to rename column from 'minus_neutral' to 'post_mpc'}
\usage{
rename_mpc(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Rename mpc
Ad hoc function to rename column from 'minus_neutral' to 'post_mpc'
}
