% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_fim}
\alias{scale_color_fim}
\title{Color scale constructor for fim colors}
\usage{
scale_color_fim(palette = "headline", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of fim_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(),
used respectively when discrete is TRUE or FALSE}
}
\value{

}
\description{
Color scale constructor for fim colors
}
