% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo_projections_calculations.R
\name{smooth_budget_series}
\alias{smooth_budget_series}
\title{Smooth budget series}
\usage{
smooth_budget_series(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Smooth budget series
}
