% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_social_benefits_components.R
\name{spread_social_benefits}
\alias{spread_social_benefits}
\title{Separate unemployment insurance and rebate checks from social benefits}
\usage{
spread_social_benefits(df)
}
\arguments{
\item{df}{}

\item{component}{}
}
\value{

}
\description{
Separate unemployment insurance and rebate checks from social benefits
}
