% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_assumptions.R
\name{state_purchases_growth_override}
\alias{state_purchases_growth_override}
\title{Override state purchases}
\usage{
state_purchases_growth_override(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Override state purchases
}
