% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_assumptions.R
\name{state_taxes}
\alias{state_taxes}
\title{State taxes}
\usage{
state_taxes(df)
}
\arguments{
\item{df}{Quarterly tax data comes from Haver.}
}
\value{

}
\description{
We assume that state and local taxes growth with GDP.
}
