% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate_assumptions.R
\name{subsidies_growth}
\alias{subsidies_growth}
\title{Subsidies growth rate assumptions}
\usage{
subsidies_growth(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Subsidies growth rate assumptions
}
