% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{sum_projections}
\alias{sum_projections}
\title{Sum federal and state projections}
\usage{
sum_projections(df, total, federal, state)
}
\arguments{
\item{df}{}

\item{total}{}

\item{federal}{}

\item{state}{}
}
\value{

}
\description{
Sum federal and state projections
}
