% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{taxes_contributions}
\alias{taxes_contributions}
\title{Apply mpcs to taxes and transfers}
\usage{
taxes_contributions(df)
}
\arguments{
\item{df}{}

\item{taxes_transfers}{}
}
\value{

}
\description{
Apply mpcs to taxes and transfers
}
