% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_helpers.R
\name{tidyselect_row_sums}
\alias{tidyselect_row_sums}
\title{Tidyselect helper to sum across rows.
Wrangle rowSums() or rowMeans() into accepting tidyverselect helpers.}
\usage{
tidyselect_row_sums(.data, ..., .value = "row_sum", na.rm = FALSE)
}
\arguments{
\item{.value}{Specify the output variable name.}

\item{na.rm}{}
}
\value{

}
\description{
Tidyselect helper to sum across rows.
Wrangle rowSums() or rowMeans() into accepting tidyverselect helpers.
}
