% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_projection.R
\name{total_projection}
\alias{total_projection}
\title{Aggregate federal and state components}
\usage{
total_projection(df, total, federal, state)
}
\arguments{
\item{df}{}

\item{total}{}

\item{federal}{}

\item{state}{}
}
\value{

}
\description{
Aggregate federal and state components
}
\examples{

   
}
