% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_projections.R
\name{total_projections}
\alias{total_projections}
\title{Aggregate federal and state components}
\usage{
total_projections(df, total, federal, state)
}
\arguments{
\item{df}{}

\item{total}{}

\item{federal}{}

\item{state}{}
}
\value{

}
\description{
Aggregate federal and state components
}
