% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{total_purchases}
\alias{total_purchases}
\title{Sum contributions of federal and state grants & purcrchases}
\usage{
total_purchases(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Sum contributions of federal and state grants & purcrchases
}
