% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocations.R
\name{unemployment_insurance_reallocation}
\alias{unemployment_insurance_reallocation}
\title{Unemployment insurance reallocation
Before constructing our projections, we reallocate unemployment insurance
from the Federal government to S&L governments-- except for emergency legislation
paid for by the Federal govt (e.g. CARES act UI expansion)}
\usage{
unemployment_insurance_reallocation(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Unemployment insurance reallocation
Before constructing our projections, we reallocate unemployment insurance
from the Federal government to S&L governments-- except for emergency legislation
paid for by the Federal govt (e.g. CARES act UI expansion)
}
