% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget-projections.R
\docType{data}
\name{budg}
\alias{budg}
\title{CBO 10-Year Budget Projections}
\format{
A data frame with 56 rows and 10 variables:
\describe{
\item{\code{date}}{Quarterly calendar data. Data was originally annual in fiscal years.}
\item{\code{gftfp}}{Federal Government Social Benefit Payments to Persons}
\item{\code{gfrpt}}{Federal Government Personal Current Tax Receipts}
\item{\code{gfrpri}}{Federal Government Tax Receipts on Production & Imports}
\item{\code{gfrcp}}{Federal Government Tax Receipts on Corporate Income}
\item{\code{gfrs}}{Contributions for Federal Government Social Insurance}
\item{\code{yptmr}}{Government Social Benefit Payments to Persons: Medicare}
\item{\code{yptmd}}{Government Social Benefit Payments to Persons: Medicaid}
\item{\code{yptu}}{Govt Transf to Persons: Unemployment Insurance Benefits}
}
}
\usage{
budg
}
\description{
Projections of spending and revenues by category and of deficits and debt held by the public.
}
\keyword{datasets}
