% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R, R/contributions.R
\name{contribution}
\alias{contribution}
\title{FUNCTION_TITLE}
\usage{
contribution(.data, ...)

contribution(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}

\item{var}{federal_nom, state_local_nom, federal_cgrants, federal_igrants}
}
\value{
Contribution of var to real gdp growth


}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{as_string}},\code{\link[rlang]{sym}}
}
