% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{counterfactual}
\alias{counterfactual}
\title{Counterfactual spending}
\usage{
counterfactual(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{

}
\description{
Counterfactual spending
}
