% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make_forecasts}
\alias{make_forecasts}
\title{Make forecast
Use growth rates to iteratively forecast}
\usage{
make_forecasts(df)
}
\arguments{
\item{df}{}

\item{forecast_period}{}

\item{components}{}
}
\value{

}
\description{
Make forecast
Use growth rates to iteratively forecast
}
