% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{neutral}
\alias{neutral}
\title{Calculate neutral counterfactual taxes and transfers}
\usage{
neutral(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculate neutral counterfactual taxes and transfers
}
