% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{taxes_transfers_minus_neutral}
\alias{taxes_transfers_minus_neutral}
\title{Subtract counterfactual taxes and transfers from realized taxes and transfers}
\usage{
taxes_transfers_minus_neutral(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Subtract counterfactual taxes and transfers from realized taxes and transfers
}
