% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocations.R
\name{unemployment_insurance_reallocation}
\alias{unemployment_insurance_reallocation}
\title{Unemployment insurance reallocation}
\usage{
unemployment_insurance_reallocation(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Unemployment insurance reallocation
}
